Allow custom eps into weno-upwind scheme
The eps parameter is used for avoiding division by zero when
computing the smoothness indicators of stencils. From Jiang and Shu
it has been proposed to set in the range of 10-5/10-7 but subsequent
studies show that in particular circumstances its value can alter
the interpolation results. It is more conventient to allow the user
to customize eps with a value eventually different from the Jiang
and Shu one.